db/requirements.txt
Django==2.2.4
django-countries==6.1.2
django-filter==2.2.0
django-loginas==0.3.6
django-modeltranslation==0.14.4
django-phonenumber-field==3.0.1
django-settings-export==1.2.1
djangorestframework==3.10.3
drf-yasg==1.17.1
flake8==3.7.8
flake8-rst-docstrings==0.0.10
ipython==7.9.0
markdown==3.1.1
phonenumbers==8.10.14
pre-commit==1.21.0
psycopg2-binary==2.8.3
python-magic==0.4.15
python-memcached==1.59
pyyaml==5.3
requests==2.19.1
selenium==3.141.0
z3c.rml==3.6.3
